Rename from duke to JeromeGpt
