+++ src/main/java/jerome/common/Messages.java
@@ -1,0 +1,12 @@
+package jerome.common;
+
+/**
+ * Represents various messages that are used in the application.
+ */
+public class Messages {
+    public static final String MESSAGE_GOODBYE = "Bye. Hope to see you again soon!";
+    // Emoji of \uD83D\uDE05 is 😅
+    public static final String MESSAGE_INCORRECT =
+            "\t I have no idea what you are trying to tell me??? \uD83D\uDE05";
+
+}
